Merge YB master commit 7fef11975a003a15d57b10d6780f01e50522e150 titled

    [PLAT-13956]: Add an RBAC action role for xCluster/DR

and committed 2024-07-26T13:25:38+00:00 into YB pg15.

- pg_dump.c:
  - dumpACL:
    - variable declarations of if (sql->len > 0): YB master
      7414711f3659350d8845bb8698480d95bd93d9f8 renames use_roles_sql to
      yb_use_roles_sql.  Upstream PG
      ea9125304dc6e90eabad165bd120eb1e667525d4 adds aclDeps and nDeps.
      Adjacent conflict.
    - ArchiveEntry call: YB master
      7414711f3659350d8845bb8698480d95bd93d9f8 does a similar rename for
      one of the arguments.  Upstream PG
      f831d4accda00b9144bc647ede2e2f848b59f39d changes the format of the
      arguments.  Merge.
- pg15_tests/passing_tests.tsv:
  - remove org.yb.pgsql.TestPgAuthorization from the list because of
    issue #23295.

+++ src/yb/yql/pggate/pg_tools.h
@@ -30,7 +30,7 @@ namespace yb::pggate {
 RowMarkType GetRowMarkType(const PgExecParameters* exec_params);
 
 struct Bound {
-  uint64_t value;
+  uint16_t value;
   bool is_inclusive;
 };
